Renamed DAO to Model to clear confusing with the meaning of DAO since the (now named) Model objects do not help accessing data from the database
