+++ Controller/SessionController.php
@@ -13,7 +13,7 @@ namespace Bundle\DoctrineUserBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\EventDispatcher\Event;
-use Bundle\DoctrineUserBundle\DAO\User;
+use Bundle\DoctrineUserBundle\Model\User;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
