+++ Document/Group.php
@@ -2,7 +2,7 @@
 
 namespace Bundle\DoctrineUserBundle\Document;
 
-use Bundle\DoctrineUserBundle\DAO\Group as AbstractGroup;
+use Bundle\DoctrineUserBundle\Model\Group as AbstractGroup;
 
 class Group extends AbstractGroup
 {
