+++ Entity/Group.php
@@ -3,7 +3,7 @@
 namespace Bundle\DoctrineUserBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Bundle\DoctrineUserBundle\DAO\Group as AbstractGroup;
+use Bundle\DoctrineUserBundle\Model\Group as AbstractGroup;
 
 class Group extends AbstractGroup
 {
