+++ Entity/Permission.php
@@ -2,7 +2,7 @@
 
 namespace Bundle\DoctrineUserBundle\Entity;
 
-use Bundle\DoctrineUserBundle\DAO\Permission as AbstractPermission;
+use Bundle\DoctrineUserBundle\Model\Permission as AbstractPermission;
 
 class Permission extends AbstractPermission
 {
