+++ Model/Group.php
@@ -1,14 +1,17 @@
 <?php
 
-namespace Bundle\DoctrineUserBundle\DAO;
+namespace Bundle\DoctrineUserBundle\Model;
+
+use Doctrine\Common\Collections\Collection;
+use Doctrine\Common\Collections\ArrayCollection;
 
 /**
- * Storage agnostic permission object
+ * Storage agnostic group object
  * Has validator annotation, but database mapping must be done in a subclass.
  *
  * @author Antoine Hérault <antoine.herault@gmail.com>
  */
-abstract class Permission
+abstract class Group
 {
     protected $id;
 
@@ -20,8 +23,11 @@ abstract class Permission
      */
     protected $name;
 
+    /**
+     * @validation:MaxLength(limit=5000)
+     */
     protected $description;
-    
+
     /**
      * @validation:DateTime()
      */
@@ -31,6 +37,11 @@ abstract class Permission
      * @validation:DateTime()
      */
     protected $updatedAt;
+
+    /**
+     * @var Collection
+     */
+    protected $permissions;
 
     public function getId()
     {
@@ -78,22 +89,6 @@ abstract class Permission
     }
 
     /**
-     * @return \DateTime
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-
-    /**
      * This method must be called just before inserting the object into the
      * database. Don't call it otherwise!
      */
@@ -112,6 +107,44 @@ abstract class Permission
     public function incrementUpdatedAt()
     {
         $this->updatedAt = new \DateTime();
+    }
+
+    /**
+     * Get permissions granted to the group 
+     * 
+     * @return Collection
+     */
+    public function getPermissions()
+    {
+        return $this->permissions ?: $this->permissions = new ArrayCollection();
+    }
+
+    /**
+     * Gets the name of the permissions granted to the group
+     *
+     * @return array
+     */
+    public function getPermissionNames()
+    {
+        $names = array();
+        foreach ($this->getPermissions() as $permission) {
+            $names[] = $permission->getName();
+        }
+
+        return $names;
+    }
+
+    /**
+     * Add a permission to the group permissions
+     *
+     * @param Permission $permission
+     * @return null
+     **/
+    public function addPermission(Permission $permission)
+    {
+        if (!$this->getPermissions()->contains($permission)) {
+            $this->getPermissions()->add($permission);
+        }
     }
 
     public function __toString()
