+++ Model/UserRepositoryInterface.php
@@ -8,7 +8,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Bundle\DoctrineUserBundle\DAO;
+namespace Bundle\DoctrineUserBundle\Model;
 
 interface UserRepositoryInterface extends RepositoryInterface
 {
